requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
h5py>=3.7.0
requests>=2.28.0
pytz>=2022.1
tzlocal>=4.2

# Web application dependencies
dash>=2.6.0
plotly>=5.10.0

# Notification dependencies
apprise>=1.9.0
python-dotenv>=0.19.0

# Development dependencies
matplotlib>=3.5.0

# YAML support
PyYAML>=5.4.0